requirements.txt
pdfminer.six==20231228
pyparsing==3.1.2